+++ src/Raven.CodeAnalysis/BoundTree/BoundConversionExpression.cs
@@ -5,17 +5,20 @@ internal partial class BoundConversionExpression : BoundExpression
     public BoundExpression Expression { get; }
     public Conversion Conversion { get; }
     public bool IsNullabilityFlowNarrowing { get; }
+    public bool IsNullableSuppression { get; }
 
     public BoundConversionExpression(
         BoundExpression expression,
         ITypeSymbol type,
         Conversion conversion,
-        bool isNullabilityFlowNarrowing = false)
+        bool isNullabilityFlowNarrowing = false,
+        bool isNullableSuppression = false)
         : base(type)
     {
         Expression = expression;
         Conversion = conversion;
         IsNullabilityFlowNarrowing = isNullabilityFlowNarrowing;
+        IsNullableSuppression = isNullableSuppression;
     }
 
     public bool IsExplicit => Conversion.IsExplicit;